SVN rev.2744~2756  - Implemented timed service without external cron  - added README.md for geoIP update  - Improved display of node status information
