+++ views/layout.pug
@@ -70,6 +70,9 @@ html(lang="en")
       hr
       footer
         span
+          a(href="https://github.com/ComBba/etherchain-light") Etherchain Light for EtherSocialNetwork
+          |&nbsp;&nbsp;&nbsp;&nbsp;
+        span
           a(href='/api') APIs service 
           |&nbsp;&nbsp;&nbsp;&nbsp;
         span
@@ -78,7 +81,6 @@ html(lang="en")
           br/ This site is operated for transparent development of "EtherSocialNetwork". 이 사이트는 "EtherSocialNetwork"의 투명한 개발을 위해 운영됩니다.
           br/ This site is currently being implemented.(status: beta test) 이 사이트는 현재 구현 중입니다.(상태: 베타 테스트).
         p
-          a(href="https://github.com/gobitfly/etherchain-light") Etherchain Light
           for VersionAndPeer in nodeStatus.VersionAndPeers
             br/
-            |   #{VersionAndPeer} Peers connected
+            | #{VersionAndPeer}
